(tests/*)+implemented tests for buffer_read_* and a random data generator
This should help with creating good mock data for tests now.  I've
forsaken the random naming of files as it creates a clutter which is
annoying to clean: better to just name the files by size.

both tests are pretty similar, using a helper function to test
differing sizes: small (1024), medium (1048576) and large (33554432).