(parser)+pres_t type ~prototype for parse_line now uses pres_t
pres_t represents a "parse result": something in between the pure
source code and the completely translated instructions.  In particular
it denotes a difference between immediate and deferred source code:
instructions that are 1-1 in with the final representation and
instructions that require an additional step.  All parse results have
a reference to the position in the source code buffer they were parsed
from, and deferred instructions usually include some critical
information for generating the final code.

process_results takes an array of parse results, the buffer of
original source code (for error message generation) and a reference to
a dynamic array to store the final representation in.  It can error
out by propagating the parse_line errors, or maybe something new.

parse_buffer should use parse_line to slurp the buffer, then
process_results to generate the final source code, which is why its
prototype remains unchanged.