(parser)~implement deferred jump for parse_line
This is a bit complex because we have 3 possible jumps in this case.

First we must ensure that an operand /does exist/ of any kind.

Absolute jumps with a negative address instantly generates a perr_t

Operands for relative jumps have a "." prefix

Label jumps check that the operand is a well formed label, then
generates a copy of it for pres_t.