(vm)~rough implementation of vm_read_program
Reads all bytes from the file then interprets data case by case, like
a really simple parser.  Working with bytes is quite nice actually.

Currently I'm using assertions for error cases, but I think we could
write some extra entries for err_t as this is part of the execution of
a vm (later on, we might add the ability to read other programs or
something?)