(data)~stupid error: should reserve 4 bits for immediate tags
Immediate tags have a 0 at the end: this allows them to be immediately
checked in comparison to future allocated types which will have a 1
bit.  Therefore I need to reserve an extra bit so that all immediate
types can be differentiated.

Funny thing is that I obviously thought about this: the max and min of
signed and unsigned integers are named 60 bit because I knew I needed
to reserve 4 bits, but I didn't define them that way because I defined
the tags this way.  Can't believe this got past me.