(parser)+introduced a new possibility for jump
Using the stack for the instruction pointer.  For the parser, this
means looking for TOKEN_HAT and setting the operand to nil (signalling
to the virtual machine that this is a stack lookup jump).  If I wanted
to implement a stack based jump which took the nth item of the stack
as an address, I could implement something similar to the relative
jump routine but use some other data type to signal to the vm the
intention.