(parser|vm)~implement instruction pointer operations
Using '*', one can refer to or use (for jumping) an instruction
pointer at runtime.  This allows users of programs to use similar
behaviour to label processing in the parser, in the sense that one
may push an instruction address and (later on) jump to it just by
using the stack.

A good use for this is creating generic routines which may have
multiple callers and must return back to the caller i.e. they cannot
halt the program after the routine is complete, and should continue
execution after that point.